improve some small things

+++ src/core/debug.h
@@ -46,7 +46,7 @@ namespace core {
 panic(const char* msg, std::source_location loc = std::source_location::current(), ...)
     PRINTF_ATTRIBUTE(1, 3);
 
-void dump_backtrace(int skip = 1);
+void dump_backtrace(usize skip = 1);
 void setup_crash_handler();
 
 } // namespace core
